refactor all buttons to use index.tsx
