+++ frontend/src/components/ResultsView/ItemMenu/PersonCard/ItemRow/itemRow.component.tsx
@@ -20,7 +20,7 @@ import {
   ERROR_INPUT_NAME,
   ERROR_INPUT_PRICE,
 } from "../../../../../utilities/variables";
-import { DeleteBtn } from "../../../../shared/buttons/deleteBtn.component";
+import { DeleteBtn } from "../../../../shared/buttons";
 import { dismissToast, errorToast } from "../../../../shared/toasts/toasts";
 
 const defaultItem: Item = {
