+++ frontend/src/components/shared/buttons/index.tsx
@@ -1,0 +1,8 @@
+export { AddBtn } from "./addBtn.component";
+export { AddPersonBtn } from "./addPersonBtn.component";
+export { CloseBtn } from "./closeBtn.component";
+export { DeleteBtn } from "./deleteBtn.component";
+export { DeletePersonBtn } from "./deletePersonBtn.component";
+export { EditBtn } from "./editBtn.component";
+export { ResetBtn } from "./resetBtn.component";
+export { ShareBtn } from "./shareBtn.component";
